move get_queryset() to common mixin
